Change the type of number from int to double
Previously, I casted every number to Cpp's int. So there was
no way to represent float in Dlox and it resulted in incorrect
values when evaluating binary and unary expressions.

To fix this, I'm casting every number to Cpp's double. This allows
me to represent both int and float by using a small trick to trim
off the leading zeros.

+++ Ast.cpp
@@ -27,8 +27,8 @@ any Ast::visitLiteralExpr(std::shared_ptr<Literal> expr) {
         } catch (const std::bad_any_cast &e) {
             std::cout << "Ast::visitLiteralExpr -> " << e.what() << "\n";
         }
-    } else if (type == typeid(int)) {
-        return std::to_string(std::any_cast<int>(expr->value));
+    } else if (type == typeid(double)) {
+        return std::to_string(std::any_cast<double>(expr->value));
     } else if (type == typeid(bool)) {
         if (std::any_cast<bool>(expr->value)) {
             std::string result{"true"};
